refactor: add new button components on pages
