+++ src/components/acessosUsuario/CadastrarUsuario.tsx
@@ -5,7 +5,7 @@ import { AuthContext } from "../../provider/Autenticacao";
 import fetchUnidadeFederativa from "../../service/unidadesFederativas";
 import fetchCadastro from "../../service/cadastrarUsuario";
 import "../../styles/form.css";
-import ButtonComponent from "../Botao";
+import { ButtonComponent } from "../Button";
 const { Option } = Select;
 
 interface UfProps {
@@ -27,7 +27,7 @@ const CadastroUsuarioForm: React.FC = () => {
   const [empresasVisiveis, setEmpresasVisiveis] = useState(false);
   const [uf, setUf] = useState<UfProps[]>();
   const empresas = [{ label: "Instituto Essência do Saber", value: 0 }];
-  const navigate = useNavigate()
+  const navigate = useNavigate();
 
   const onFinish = async (values: any) => {
     const cadastroUsuarioData = {
@@ -40,7 +40,7 @@ const CadastroUsuarioForm: React.FC = () => {
     try {
       await fetchCadastro(cadastroUsuarioData);
       notification.success({ message: "Cadastro feito!" });
-      navigate('/login')
+      navigate("/login");
     } catch (error) {
       api.error({ message: `Erro ao fazer cadastro` });
     }
@@ -223,11 +223,9 @@ const CadastroUsuarioForm: React.FC = () => {
           <Form.Item>
             <Space>
               <ButtonComponent
-                nome="Cadastrar-se"
-                cor="#1351B4"
-                cor_letra="#FFFFFF"
-                cor_borda="#1351B4"
-                largura="10em"
+                label="Cadastrar-se"
+                buttonStyle="primary"
+                padding="40px"
               />
             </Space>
           </Form.Item>
