+++ src/components/acessosUsuario/Login.tsx
@@ -5,10 +5,9 @@ import LogoDNIT from "../../assets/logoDnitAzul.png";
 import { AuthContext } from "../../provider/Autenticacao";
 import fetchLogin from "../../service/login";
 import "../../styles/form.css";
-import ButtonComponent from "../Botao";
+import { ButtonComponent } from "../Button";
 
 const LoginForm: React.FC = () => {
-
   const [form] = Form.useForm();
   const regras = [
     {
@@ -35,7 +34,7 @@ const LoginForm: React.FC = () => {
 
     try {
       await fetchLogin(loginData);
-      notification.success({message: "Login realizado!"})
+      notification.success({ message: "Login realizado!" });
       login();
     } catch (error) {
       api.error({ message: `Erro ao fazer login` });
@@ -75,11 +74,9 @@ const LoginForm: React.FC = () => {
           <Form.Item>
             <Space>
               <ButtonComponent
-                nome="Entrar"
-                cor="#1351B4"
-                cor_letra="#FFFFFF"
-                cor_borda="#1351B4"
-                largura="10em"
+                label="Entrar"
+                buttonStyle="primary"
+                padding="40px"
               />
             </Space>
           </Form.Item>
