+++ src/components/recuperarSenha/RecuperarSenha.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import LogoDNIT from "../../assets/logoDnitAzul.png";
 import fetchRecuperarSenha from "../../service/recuperarSenha";
 import "../../styles/form.css";
-import ButtonComponent from "../Botao";
+import { ButtonComponent } from "../Button";
 
 const EsqueciSenhaForm: React.FC = () => {
   const [formulario] = Form.useForm();
@@ -55,11 +55,9 @@ const EsqueciSenhaForm: React.FC = () => {
           <Form.Item className="esqueci">
             <Space>
               <ButtonComponent
-                nome="Enviar link de recuperação"
-                cor="#1351B4"
-                cor_letra="#FFFFFF"
-                cor_borda="#1351B4"
-                largura="16em"
+                label="Enviar Link de Recuperação"
+                buttonStyle="primary"
+                padding="40px"
               />
             </Space>
           </Form.Item>
