+++ src/components/recuperarSenha/RedefinirSenha.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import LogoDNIT from "../../assets/logoDnitAzul.png";
 import fetchRedefinirSenha from "../../service/redefinirSenha";
 import "../../styles/form.css";
-import ButtonComponent from "../Botao";
+import { ButtonComponent } from "../Button";
 
 const RedefinirSenha: React.FC = () => {
   const [formulario] = Form.useForm();
@@ -85,11 +85,9 @@ const RedefinirSenha: React.FC = () => {
           <Form.Item className="recuperarsenha">
             <Space>
               <ButtonComponent
-                nome="Confirmar"
-                cor="#1351B4"
-                cor_letra="#FFFFFF"
-                cor_borda="#1351B4"
-                largura="10em"
+                label="Confirmar"
+                buttonStyle="primary"
+                padding="40px"
               />
             </Space>
           </Form.Item>
